+++ NEWS.md
@@ -1,5 +1,11 @@
 # omopgenerics 1.1.0
 
+* more general cdm validation checks by @edward-burn in #674
+* typo in validateConceptSet by @catalamarti in #673
+* fix call argument by @catalamarti in #677
+* fix tempdir("...") by @catalamarti in #679
+* new function transformToSummarisedResult by @catalamarti in #676
+
 # omopgenerics 1.0.0
 
 * Stable release of the package.